refactor(mobile): generalize double-tap detection into useDoubleTap<Key>
The gesture is generic — two taps of the same key within a window,
expired when the active key moves off it — so the hook drops its
MobileTab/Route coupling and moves to hooks/. tabRootFor lands beside
MobileTab in mobile-tab-bar; the shell composes the two.
